agents/perception/pyproject.toml
[project]
name = "chimera-perception"
version = "1.0.0"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115",
  "uvicorn[standard]>=0.30",
  "pillow>=10.0",
  "mss>=9.0",
  "pytesseract>=0.3.10",
  "httpx>=0.27",
  "python-dotenv>=1.0",
]
